Rename Table to Schema for consistency and less confusion.
